Updated to make types match after moving concept to uint32_t

+++ contracts/include/reward-series.hpp
@@ -20,7 +20,7 @@
 
 
     // --- Return a Series --- //
-    vector<uint64_t> getSeries(const string& seriesName){
+    vector<uint32_t> getSeries(const string& seriesName){
         if (seriesName == "FIBONACCI") {
             return series::FIBONACCI;
         } else if (seriesName == "SILVER") {
